Separate userRouter
